+++ spec2vec/Spec2Vec.py
@@ -80,7 +80,7 @@ class Spec2Vec(BaseSimilarity):
 
     """
     def __init__(self, model: Word2Vec, intensity_weighting_power: Union[float, int] = 0,
-                 allowed_missing_percentage: Union[float, int] = 50, progress_bar: bool = False):
+                 allowed_missing_percentage: Union[float, int] = 10, progress_bar: bool = False):
         """
 
         Parameters
@@ -95,8 +95,8 @@ class Spec2Vec(BaseSimilarity):
         allowed_missing_percentage:
             Set the maximum allowed percentage of the document that may be missing
             from the input model. This is measured as percentage of the weighted, missing
-            words compared to all word vectors of the document. Default is 100, which
-            means up to 50% missing words are allowed (not a very strict setting!).
+            words compared to all word vectors of the document. Default is 10, which
+            means up to 10% missing words are allowed before a warning is raised.
         progress_bar:
             Set to True to monitor the embedding creating with a progress bar.
             Default is False.